AAE-7173 abstract tests
