+++ activiti-cloud-modeling-service/activiti-cloud-services-modeling/activiti-cloud-services-modeling-rest/pom.xml
@@ -134,5 +134,10 @@
       <artifactId>activiti-cloud-services-test-containers</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>activiti-cloud-services-test-security-keycloak</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
